+++ src/index.ts
@@ -5,5 +5,6 @@ export * from './misc/enums';
 export * from './misc/providers';
 export * from './misc/tokens';
 export * from './misc/types';
+export * from './misc/utils';
 export * from './modules';
 export * from './pipes';