+++ src/plugins/interactions/noInteractions/noInteractions.component.ts
@@ -21,7 +21,7 @@ export class NoInteractionsComponent implements Interactions
     //######################### constructor #########################
     constructor(public pluginElement: ElementRef<HTMLElement>,
                 public selectPlugins: SelectPluginInstances,
-                public pluginBus: SelectBus<unknown>,)
+                public selectBus: SelectBus<unknown>,)
     {
 
     }