+++ src/plugins/liveSearch/filterLiveSearch/filterLiveSearch.component.ts
@@ -21,7 +21,7 @@ export class FilterLiveSearchComponent implements LiveSearch
     //######################### constructor #########################
     constructor(public pluginElement: ElementRef<HTMLElement>,
                 public selectPlugins: SelectPluginInstances,
-                public pluginBus: SelectBus<unknown>,)
+                public selectBus: SelectBus<unknown>,)
     {
 
     }